Use name and version from pkgdef' in builder
Previously there was a mix of using `pname` and `version` from the nix
derivation attributes, using `name` and `version` from the *original*
pkgdef, and using `name` and `version` from the final overridden
(`pkgdef'`). This can lead to a situation where there is inconsistency
in different usages within the same package.

In particular, this fixes a problem where a matched dependency (i.e. a
dependency with `{ = version }` constraint) of a prerelease versions
(e.g., `1.0.0~beta1`) would be missing since dependency filtering was
using the *nix* version attribute (which replaces '~' by '_') not the
*opam* version attribute. This caused the correctly resolved dependency
to be filtered out.